System structure change and error handling
LSystem class has been modified to use two hashtables. The terminal
hashtable, removes the need for the two grammar slots and the angle
slot. It farther changes the rule structure to allow for different
angles to be defined.

A few error handling functions have been defined to provide checks
for missing data.

In create-object, the top let binding has been removed, and the
bindings have been included into the iterate fuction; as well as,
the setting of the origin point. The handling of the terminals has
also been modified to better utilize the new hashtable.

In substitution, append has been replaced with format, fixing a
crash that occurs on certain n values.

The previous unit tests have been re-written to work with the recent
changes.

The data folder has been changed from res to data. README has been
updated to reflect changes.